Updating PHP Support Versions
The table of supported PHP versions was out-of-date, so I updated it with the current available versions and marking EOL for versions no longer supported by the Core PHP crew.
I am not sure yet if PHP 8.1 will be supported on the Linux only platform since I haven't seen any communication about it. It was released November 25, 2021.